+++ client/src/main/java/replicant/WebSocketConfig.java
@@ -5,6 +5,8 @@ import java.util.function.Consumer;
 import java.util.function.Supplier;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
+import org.realityforge.gwt.keycloak.Keycloak;
+import org.realityforge.replicant.shared.SharedConstants;
 
 public final class WebSocketConfig
 {
@@ -18,7 +20,7 @@ public final class WebSocketConfig
    * This runnable may be invoked immediately or deferred until later (such as after
    * authentication token has been generated).
    */
-  @Nonnull
+  @Nullable
   private final Consumer<Runnable> _remoteCallWrapper;
   /**
    * Mechanism via which an authentication token is generated, if any.
@@ -26,12 +28,26 @@ public final class WebSocketConfig
   @Nullable
   private final Supplier<String> _authenticationTokenGenerator;
 
-  public WebSocketConfig( @Nonnull final String url,
-                          @Nonnull final Consumer<Runnable> remoteCallWrapper,
-                          @Nullable final Supplier<String> authenticationTokenGenerator )
+  @Nonnull
+  public static WebSocketConfig create( @Nonnull final String baseURL )
+  {
+    return new WebSocketConfig( toWebSocketProtocol( baseURL ) + SharedConstants.REPLICANT_URL_FRAGMENT, null, null );
+  }
+
+  @Nonnull
+  public static WebSocketConfig create( @Nonnull final String baseURL, @Nonnull final Keycloak keycloak )
+  {
+    return new WebSocketConfig( toWebSocketProtocol( baseURL ) + SharedConstants.REPLICANT_URL_FRAGMENT,
+                                keycloak::updateTokenAndExecute,
+                                keycloak::getToken );
+  }
+
+  private WebSocketConfig( @Nonnull final String url,
+                           @Nullable final Consumer<Runnable> remoteCallWrapper,
+                           @Nullable final Supplier<String> authenticationTokenGenerator )
   {
     _url = Objects.requireNonNull( url );
-    _remoteCallWrapper = Objects.requireNonNull( remoteCallWrapper );
+    _remoteCallWrapper = remoteCallWrapper;
     _authenticationTokenGenerator = authenticationTokenGenerator;
   }
 
@@ -49,6 +65,19 @@ public final class WebSocketConfig
 
   public void remote( @Nonnull final Runnable action )
   {
-    _remoteCallWrapper.accept( action );
+    if ( null == _remoteCallWrapper )
+    {
+      action.run();
+    }
+    else
+    {
+      _remoteCallWrapper.accept( action );
+    }
+  }
+
+  @Nonnull
+  private static String toWebSocketProtocol( @Nonnull final String baseURL )
+  {
+    return baseURL.replace( "https://", "wss://" ).replace( "http://", "ws://" );
   }
 }