+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+### Unreleased
+
 ### [v6.72](https://github.com/replicant4j/replicant/tree/v6.72) (2020-02-20) · [Full Changelog](https://github.com/replicant4j/replicant/compare/v6.71...v6.72)
 
 * Refactor `ReplicantRpcRequestBuilder` to accept the base url during construction.